backend: lobby router and controller
