+++ backend/src/api/star-systems/starSystems.router.test.ts
@@ -3,7 +3,7 @@ import { v4 } from "uuid"
 import { describe, expect, it } from "vitest"
 import { createNewAccountModelStub } from "#api/accounts/NewAccountModel.stub.ts"
 import { createApiStub } from "#api/createApi.stub.ts"
-import { createCreateGameDtoStub } from "#api/games/CreateGameDto.stub.ts"
+import { createCreateGameDtoStub } from "#api/game-lobbies/CreateGameDto.stub.ts"
 import type { GameId } from "#api/games/GameId.ts"
 import { type NewStarSystemModel, type StarSystemsRepository } from "#lib/db/star-systems/starSystems.repository.ts"
 import { BodyType } from "#lib/star-systems/BodyType.ts"
@@ -19,7 +19,7 @@ describe("starSystems.router", () => {
 
       authService.account = extractSuccess(await accountsRepository.createAccount(createNewAccountModelStub()))
 
-      const { createdGameId } = await trpcClient.client.games.create.mutate(createCreateGameDtoStub())
+      const { createdGameId } = await trpcClient.client.gameLobbies.create.mutate(createCreateGameDtoStub())
       const starSystem = await createStoredStarSystem({ starSystemsRepository, gameId: createdGameId })
 
       // Act
@@ -137,7 +137,7 @@ describe("starSystems.router", () => {
       const outsider = extractSuccess(await accountsRepository.createAccount(createNewAccountModelStub({ alias: "Outsider" })))
       authService.account = creator
 
-      const { createdGameId } = await trpcClient.client.games.create.mutate(createCreateGameDtoStub())
+      const { createdGameId } = await trpcClient.client.gameLobbies.create.mutate(createCreateGameDtoStub())
       await createStoredStarSystem({ starSystemsRepository, gameId: createdGameId })
       authService.account = outsider
 
@@ -154,7 +154,7 @@ describe("starSystems.router", () => {
 
       authService.account = extractSuccess(await accountsRepository.createAccount(createNewAccountModelStub()))
 
-      const { createdGameId } = await trpcClient.client.games.create.mutate(createCreateGameDtoStub())
+      const { createdGameId } = await trpcClient.client.gameLobbies.create.mutate(createCreateGameDtoStub())
 
       // Act & Assert
       await expect(trpcClient.client.starSystems.getByGameId.query({ gameId: createdGameId })).rejects.toMatchObject({
